+++ src/main/java/subway/dao/SectionDao.java
@@ -1,90 +1,27 @@
 package subway.dao;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import org.springframework.dao.DataAccessException;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 import subway.dao.entity.SectionEntity;
 import subway.domain.Section;
-import subway.domain.Station;
 
-@Repository
-public class SectionDao {
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert insertAction;
-    private static final RowMapper<SectionEntity> rowMapper = (rs, rowNum) -> new SectionEntity(
-            rs.getLong("id"),
-            rs.getLong("line_id"),
-            new Station(rs.getString("start_station_name")),
-            new Station(rs.getString("end_station_name")),
-            rs.getInt("distance"));
+public interface SectionDao {
 
-    public SectionDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.insertAction = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("section")
-                .usingGeneratedKeyColumns("id");
-    }
+    long insert(Long lineId, Section section);
 
-    public long insert(Long lineId, Section section) {
-        return insertAction.executeAndReturnKey(Map.of(
-                "line_id", lineId,
-                "start_station_name", section.getUpBoundStation().getName(),
-                "end_station_name", section.getDownBoundStation().getName(),
-                "distance", section.getDistance()
-        )).longValue();
-    }
+    void update(Long sectionId, Section section);
 
-    public void update(Long sectionId, Section section) {
-        String sql = "UPDATE SECTION SET start_station_name = ?, end_station_name = ?, distance = ? WHERE id = ?";
-        jdbcTemplate.update(sql, section.getUpBoundStationName(), section.getDownBoundStationName(),
-                section.getDistance(), sectionId);
-    }
+    Long countByLineId(Long lineId);
 
-    public Long countByLineId(Long lineId) {
-        String sql = "SELECT count(*) FROM SECTION WHERE line_id = ?";
-        return jdbcTemplate.queryForObject(sql, Long.class, lineId);
-    }
+    List<SectionEntity> findAllByLineId(Long lineId);
 
-    public List<SectionEntity> findAllByLineId(Long lineId) {
-        String sql = "SELECT id, line_id, start_station_name, end_station_name, distance FROM SECTION WHERE line_id = ?";
-        return jdbcTemplate.query(sql, rowMapper, lineId);
-    }
+    boolean isStationInLine(Long lineId, String stationName);
 
-    public boolean isStationInLine(Long lineId, String stationName) {
-        String sql = "SELECT COUNT(*) FROM SECTION WHERE start_station_name = ? OR end_station_name = ? AND line_id = ? LIMIT 1";
-        return jdbcTemplate.queryForObject(sql, Long.class, stationName, stationName, lineId) > 0;
-    }
+    boolean isEmptyByLineId(Long lineId);
 
-    public boolean isEmptyByLineId(Long lineId) {
-        String sql = "SELECT COUNT(*) FROM SECTION WHERE line_id = ? LIMIT 1";
-        return jdbcTemplate.queryForObject(sql, Long.class, lineId) == 0;
-    }
+    Optional<SectionEntity> findByStartStationNameAndLineId(String upBoundStationName, Long lineId);
 
-    public Optional<SectionEntity> findByStartStationNameAndLineId(String startStationName, Long lineId) {
-        String sql = "SELECT id, line_id, start_station_name, end_station_name, distance FROM SECTION WHERE start_station_name = ? AND line_id = ?";
-        try {
-            return Optional.ofNullable(jdbcTemplate.queryForObject(sql, rowMapper, startStationName, lineId));
-        } catch (DataAccessException e) {
-            return Optional.empty();
-        }
-    }
+    Optional<SectionEntity> findByEndStationNameAndLineId(String downBoundStationName, Long lineId);
 
-    public Optional<SectionEntity> findByEndStationNameAndLineId(String endStationName, Long lineId) {
-        String sql = "SELECT id, line_id, start_station_name, end_station_name, distance FROM SECTION WHERE end_station_name = ? AND line_id = ?";
-        try {
-            return Optional.ofNullable(jdbcTemplate.queryForObject(sql, rowMapper, endStationName, lineId));
-        } catch (DataAccessException e) {
-            return Optional.empty();
-        }
-    }
-
-    public void deleteBy(SectionEntity sectionEntity) {
-        String sql = "DELETE FROM SECTION WHERE id = ?";
-        jdbcTemplate.update(sql, sectionEntity.getId());
-    }
+    void deleteBy(SectionEntity sectionEntity);
 }
